cor do sucesso

+++ bin/apc.js
@@ -1,4 +1,7 @@
 #! /usr/bin/env node
+var trace = require("./ConsoleColors");
+var ProgressBar = require('progress');
+
 var shell = require("shelljs");
 
 var walk = require('fs-walk');
@@ -12,7 +15,7 @@ var uppercamelcase = require('uppercamelcase');
 const argv = require('minimist')(process.argv.slice(2));
 const FILENAME = process.mainModule.filename;
 
-const mensagemUso = "Uso: apc --c=comando -n=nome-do-modulo [--src=src do projeto] [--modules=diretorio de módulos do projeto]";
+const mensagemUso = "Uso: apc --c=comando -n=nome-do-modulo [--add Adiciona aos modulos] [--src=src do projeto] [--modules=diretorio de módulos do projeto]";
 
 var pathPartes = FILENAME.split('\\');
 pathPartes.pop();
@@ -34,17 +37,17 @@ const DEBUG = false;
 if (fse.existsSync('package.json')) {
     var projectPackage = JSON.parse(fse.readFileSync('package.json', 'utf8'));
     if (!projectPackage.dependencies['@angular/common'] || DEBUG) {
-        console.log("Não é um projeto Angular");
+        trace.error("Não é um projeto Angular");
         if (!DEBUG) {
             process.exit(1);
         }
     }
 } else {
+    trace.error("Não existe package.json");
     if (!DEBUG) {
         process.exit(2);
     }
 }
-
 function removerAcentos( newStringComAcento ) {
     var string = newStringComAcento;
     var mapaAcentosHex 	= {
@@ -91,7 +94,15 @@ function refactorRename(name) {
 }
 
 function proccessRefactorRename(path, dictNames) {
-
+    /*
+    var bar = new ProgressBar('  Criando Módulo [:bar] :percent', {
+        complete: '=',
+        incomplete: ' ',
+        width: 20,
+        total: 100
+    });
+    var contador = 0;
+    */
     walk.walkSync(path, function(basedir, filename, stat) {
 
         if (!stat.isDirectory()) {
@@ -108,37 +119,68 @@ function proccessRefactorRename(path, dictNames) {
             for(var n in dictNames) {
                 content = content.split(n).join(dictNames[n]);
             }
-            /*
-            content = content.split("clear-template").join(dictNames["clear-template"]);
-            content = content.split("ClearTemplate").join(dictNames["ClearTemplate"]);
-            content = content.split("clearTemplate").join(dictNames["clearTemplate"]);
-            content = content.split("CLEAR_TEMPLATE").join(dictNames["CLEAR_TEMPLATE"]);
-            */
-            console.log(newFile);
+
             fs.renameSync (file, newFile);
             fs.writeFileSync(newFile, content, { encoding: "utf8" });
+
+            //bar.tick(contador++);
         }
     });
 
-    console.log("DONE proccessRefactorRename");
+    trace.msg("Módulo criado com sucesso.", trace.FgGreen);
+}
+
+function newModule() {
+    if (!argv.n) {
+        trace.warning("Warging: Nome do módulo não definido");
+        trace.warning(mensagemUso);
+        process.exit(3);
+    }
+    var dictNames = refactorRename(argv.n);
+
+    var newModulePath = PROJECT_MODULES + "\\" + dictNames['clear-template'];
+    if (fs.existsSync(newModulePath)) {
+        trace.error("Erro: já existe um módulo com o mesmo nome.\nA operação não foi realizada.");
+        process.exit(1);
+    }
+    if (newModulePath.charAt(newModulePath.length - 1) != "\\") {
+        newModulePath += "\\";
+    }
+    fse.copySync(BIN_PATH + "\\templates\\module\\clear-template", newModulePath);
+    proccessRefactorRename(newModulePath, dictNames);
+
+    var modulosConstantes = PROJECT_MODULES +  "/gerenciar-modulos/shared/constants/modulos.constants.ts";
+    var modulosConstantesContent = fs.readFileSync(modulosConstantes, "utf8");
+
+    if (argv.add) {
+        var novoModulo = '\tpublic static CLEAR_TEMPLATE: string = "CLEAR_TEMPLATE";';
+        novoModulo = novoModulo.split("CLEAR_TEMPLATE").join(dictNames["CLEAR_TEMPLATE"]);
+        modulosConstantesContent = modulosConstantesContent.split("}").join(novoModulo + "\n\n}");
+        fs.writeFileSync(modulosConstantes, modulosConstantesContent, { encoding: "utf8" });
+    }
+}
+
+function novoProjeto(projectName) {
+    if (!argv.n) {
+        trace.warning("Warging: Nome do projeto não definido");
+        trace.warning(mensagemUso);
+        process.exit(3);
+    }
+
+    shell.exec("ng new " + projectName);
+    shell.exec("cd " + projectName);
 }
 
 switch (argv.c) {
 
+    case "new": {
+        novoProjeto(argv.n);
+        break;
+    }
+
     case "m":
     case "module": {
-        if (!argv.n) {
-            console.log("Warging: Nome do módulo não definido");
-            console.log(mensagemUso);
-            process.exit(3);
-        }
-        var dictNames = refactorRename(argv.n);
-        var newModulePath = PROJECT_MODULES + "\\" + dictNames['clear-template'];
-        if (newModulePath.charAt(newModulePath.length - 1) != "\\") {
-            newModulePath += "\\";
-        }
-        fse.copySync(BIN_PATH + "\\templates\\module\\clear-template", newModulePath);
-        proccessRefactorRename(newModulePath, dictNames);
+        newModule();
         break;
     }
